Fix lib_dir
The lib_dir is the first argument in a data_files pair. The
documentation for this is confusing, but the first argument should be a
directory name to be used when installing (not a currently existing
directory), and will be interpreted relative to a prefix such as
site.USER_BASE. So site.USER_BASE shouldn't be included here. Typically
this would be a single flat directory, so "libmecab" is fine.