Add the 26 field UniDic format
There are four 2.1.2 distribution files:

    unidic-mecab-2.1.2_src.zip
    unidic-mecab-2.1.2_bin.zip
    unidic-mecab-2.1.2_model.zip
    unidic-mecab_kana-accent-2.1.2_src.zip

The `src` distribution requires compilation with MeCab. The `model`
distribution is for reference. The `bin` distribution doesn't have a
note explaining it, but is actually based on the `kana-accent` file, not
the `src` file.

This change adds support for the format used in `bin`, since that's
often the easiest relatively modern form of UniDic to install.