Add support for fields containing commas
In Unidic, accent data can have a literal field value like "0,2". This
means that just splitting the features on commas won't work.
Additionally, there's no way to change the delimiter MeCab uses, so it's
necessary to do a proper parse of the line.

This change is sure to work, but it may make feature generation slower -
that will need to be checked eventually.