Use -C option in MeCab
Using this option causes MeCab to make a copy of input strings, which
means that when getting Nodes the surface doesn't have to be copied out
immediately.

In the usual benchmark this is marginally faster than the current way of
doing things, but only barely (<1%). Not sure if it should be merged to
master.